==> 1_Retrieve Application Numbers.ipynb <==
from pathlib import Path
import pandas as pd
import numpy as np
import os
import requests
from requests.auth import HTTPBasicAuth
import time
from google.colab import files

# 1. Configure API to EPO OPS

Registro en OPS para la API: [OPS](https://www.epo.org/en/searching-for-patents/data/web-services/ops)

Documentacion de la API: [OPS v1.3.19](https://link.epo.org/web/ops_v3.2_documentation_-_version_1.3.19_en.pdf)

client_id = 'i6AxJ1MYJ43N3SkC5CeRpujtzndUNHzRk72kulUA04n4wh8M'
client_secret = 'SFmlAli4KOlr1xPNygxRfgXNcfUmDzrtjpAkfv10Hkf6R1cEQsmXCBZjXV5zyhaQ'

# Authentication URL for the EPO API
auth_url = 'https://ops.epo.org/3.2/auth/accesstoken'

# Send a POST request to get the access token
auth_response = requests.post(auth_url,
                              auth=HTTPBasicAuth(client_id, client_secret),
                              headers={'Content-Type': 'application/x-www-form-urlencoded'},
                              data={'grant_type': 'client_credentials'})

# Parse the JSON response
auth_data = auth_response.json()

# Extract the access token from the response
access_token = auth_data['access_token']

# URL for the EPO API search endpoint
search_url = 'https://ops.epo.org/3.2/rest-services/published-data/search/biblio'

def get_patent_data(query_cql, start, end):
    """
    Function to retrieve patent data from the EPO API.

    :param query_cql: The CQL query string for searching patents
    :param start: The start index of the range of results to retrieve
    :param end: The end index of the range of results to retrieve
    :return: JSON response with patent data or None if there's an error
    """
    # Set up the headers for the API request
    headers = {
        'Authorization': f'Bearer {access_token}',
        'Accept': 'application/json',
        'Range': f'{start}-{end}'  # Specify the range of results to retrieve
    }

    # Set up the query parameters
    params = {
        'q': query_cql
    }

    # Send a GET request to the EPO API
    response = requests.get(search_url, headers=headers, params=params)

    # Check if the request was successful
    if response.status_code == 200:
        return response.json()
    else:
        print(f'Error: {response.status_code} - {response.text}')
        return None

# Example CQL query to search for patents in the A61K IPC class
query_cql = 'ipc=/low A61K'

# Maximum number of results allowed per request by the API
batch_size = 100

data = get_patent_data(query_cql, start=1, end=3)

data

docs = data['ops:world-patent-data']['ops:biblio-search'][ 'ops:search-result']['exchange-documents']
[
    {
        'family' : doc['exchange-document']['@family-id'],
        'publication_number' : doc['exchange-document']['bibliographic-data']['publication-reference']['document-id'][0]['doc-number']['$'],
        'application_number' : doc['exchange-document']['bibliographic-data']['application-reference']['document-id'][0]['doc-number']['$']
    } for doc in docs
]

docs[0]

#### Loop to retrieve all files

start = 1
doc_list = []
while True:

    end = start + batch_size - 1
    data = get_patent_data(query_cql, start, end)
    if not data:
        break
    docs = data['ops:world-patent-data']['ops:biblio-search'][ 'ops:search-result']['exchange-documents']
    doc_list = doc_list + [
            {
            'family' : doc['exchange-document']['@family-id'],
            'publication_number' : doc['exchange-document']['bibliographic-data']['publication-reference']['document-id'][0]['doc-number']['$'],
            'application_number' : doc['exchange-document']['bibliographic-data']['application-reference']['document-id'][0]['doc-number']['$']
            } for doc in docs
    ]
    time.sleep(3) # wait few seconds not to get blocked
    start = end

len(doc_list)

end

pd.DataFrame(doc_list).drop_duplicates(ignore_index=True).shape

df = pd.DataFrame(doc_list).drop_duplicates(ignore_index=True)
df.to_csv('Patents_publication_numbers.csv', index=False)

files.download('Patents_publication_numbers.csv') #Download from Google Colab

Para sacar los datos hicimos este paso para todas las IPC que se nos dieron en un principio:<br>
C07D, A61P, A61K, C07B, C07C, C07K, C07F, C07H, C12N, G01N, B12N (biosimilars), for industry comparison: telecommunication (Apple, Samsung): G06F, H04W, H01L <br>

Depues exportamos cada uno y los juntamos todos a mano.
