==> 2_Bajar_HTMLS.ipynb <==
import pandas as pd
import numpy as np
import re
from bs4 import BeautifulSoup
import requests
import json
import time
import random
from ast import literal_eval
import os
import zipfile
from google.colab import files

# Get lapto's headers: https://httpbin.org/anything
headers =  {
    "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7",
    #"Accept-Encoding": "gzip, deflate, br, zstd",
    "Accept-Language": "en-US,en;q=0.9,es;q=0.8",
    "Dnt": "1",
    #"Host": "httpbin.org",
    "Referer": "https://www.scraperapi.com/",
    "Sec-Ch-Ua": "\"Google Chrome\";v=\"125\", \"Chromium\";v=\"125\", \"Not.A/Brand\";v=\"24\"",
    "Sec-Ch-Ua-Mobile": "?0",
    "Sec-Ch-Ua-Platform": "\"Windows\"",
    "Sec-Fetch-Dest": "document",
    "Sec-Fetch-Mode": "navigate",
    "Sec-Fetch-Site": "cross-site",
    "Sec-Fetch-User": "?1",
    "Upgrade-Insecure-Requests": "1",
    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/125.0.0.0 Safari/537.36",
    "X-Amzn-Trace-Id": "Root=1-665ec732-28d2172a4721ed726fc59ceb"
  }

def get_page(an):
    URL = f"https://register.epo.org/smartSearch?query={an}" #https://register.epo.org/smartSearch?query= #https://register.epo.org/application?number=EP
    page = requests.get(URL,headers=headers)
    return page.text

def get_redirected_url(an): #Because of patent family info is in other tab we have to build the url to scrap
    initial_url = f"https://register.epo.org/smartSearch?query={an}"
    response = requests.get(initial_url, headers=headers, allow_redirects=True)
    redirected_url = response.url
    return redirected_url

def get_family_tab_content(an):# Scrap the patent family info
    redirected_url = get_redirected_url(an)
    modified_url = f"{redirected_url}&lng=en&tab=family"
    response = requests.get(modified_url, headers=headers)
    return response.text

application_number = "EP4215191"
redirected_url = get_redirected_url(application_number)
modified_url = f"{redirected_url}&lng=en&tab=family"
modified_url

page_content = get_family_tab_content(application_number)
filename = f'HTMLS/{application_number}_Family_Tab.txt'
with open(filename, 'w', encoding='utf-8') as file:
  file.write(page_content)

df = pd.read_csv('Patents_publication_numbers.csv')
for publication_number in df['publication_number']:
    try:
      print("Procesando: "+publication_number)
      page_content = get_page(publication_number)
      filename = f'HTMLS/{publication_number}.txt'
      with open(filename, 'w', encoding='utf-8') as file:
          file.write(page_content)
    except Exception as e:
            print(f"Error inesperado al procesar el código {publication_number}: {str(e)}")
            continue
    time.sleep(15)


df = pd.read_csv('Patents_publication_numbers.csv')
for publication_number in df['publication_number']:
    try:
      print("Procesando: "+publication_number)
      page_content = get_family_tab_content(publication_number)
      filename = f'HTMLS/Patents_Family/{publication_number}_Patent_Family.txt'
      with open(filename, 'w', encoding='utf-8') as file:
          file.write(page_content)
    except Exception as e:
            print(f"Error inesperado al procesar el código {publication_number}: {str(e)}")
            continue
    time.sleep(15)

folder_name = 'HTMLS'  # Folder in Colab where we are saving the data

zip_filename = f'{folder_name}.zip'

# Compress the data and download it
def zip_folder(folder_name, zip_filename):
    with zipfile.ZipFile(zip_filename, 'w', zipfile.ZIP_DEFLATED) as zipf:
        for root, _, files in os.walk(folder_name):
            for file in files:
                zipf.write(os.path.join(root, file))

zip_folder(folder_name, zip_filename)

files.download(zip_filename)

print(f"La carpeta {folder_name} ha sido comprimida y está lista para descargar.")
