==> 8_llm.ipynb <==
# %pip install llama-index-embeddings-huggingface
# %pip install llama-index-vector-stores-chroma
# %pip install llama-index chromadb --quiet
# %pip install llama-index-llms-langchain
# %pip install sentence-transformers
# %pip install langchain_community
# %pip install pydantic==1.10.11
# %pip install llama-index
# %pip install nbconvert
# %pip install chromadb

# import
from llama_index.core import VectorStoreIndex, SimpleDirectoryReader
from llama_index.vector_stores.chroma import ChromaVectorStore
from llama_index.core import StorageContext
from llama_index.embeddings.huggingface import HuggingFaceEmbedding
from IPython.display import Markdown, display
import chromadb
from langchain_community.llms import Ollama
from llama_index.core import Settings

# create client and a new collection
chroma_client = chromadb.EphemeralClient() #datos que solo existen en memoria temporalmente
chroma_collection = chroma_client.create_collection("prueba2")

# define embedding function
embed_model = HuggingFaceEmbedding(model_name="sentence-transformers/paraphrase-mpnet-base-v2")

# load documents
documents = SimpleDirectoryReader("/Users/diego/Desktop/rag2").load_data()

# set up ChromaVectorStore and load in data
vector_store = ChromaVectorStore(chroma_collection=chroma_collection)
storage_context = StorageContext.from_defaults(vector_store=vector_store)
index = VectorStoreIndex.from_documents(
    documents, storage_context=storage_context, embed_model=embed_model
)

# Initialize a LLM using Ollama
llm = Ollama(model="llama3.1:8b-instruct-q2_K")

# Set the LLM instance in the global configuration class 'Settings'
Settings.llm = llm

# Assign the previously defined embedding model ('embed_model') to the global configuration 'Settings'
Settings.embed_model = embed_model


# Query Data from the persisted index
#Pregunta a la patente abuelo
query_engine = index.as_query_engine()
response = query_engine.query("Please analyze and understand the following question before providing a response: What are the divisional application numbers associated with patent EP1492505? I am looking for all the divisional applications that have been filed under this main patent.")
print(response)

# Query Data from the persisted index
# Pregunta a la patente padre
query_engine = index.as_query_engine()

response = query_engine.query("Please analyze and understand the following question before providing a response: For patent EP2425824, can you provide the parent application numbers and any divisional application numbers associated with this patent? I am looking for all the parent and divisional applications linked to this patent, if they exist.")
print(response)

# Query Data from the persisted index
# Pregunta a la patente nieto
query_engine = index.as_query_engine()
response = query_engine.query("Please analyze and understand the following question before providing a response: What is the current status of patent EP3326618? I am looking for detailed information on whether this patent is active, expired, pending, or in any other status.")
print(response)
